Performance: Don't copy the input bytes as string to the PhpToken.
This reduces memory allocation for doubles, bools and floats, as they can be directly evaluated from the raw bytestream.  That also speeds up deserialization by a lot.

We also no longer have to give the Tokenizer the input encoding.

There are probably some further optimizations to be had in the deserializer -> DeserializeTokenFromSimpleType() method too.