Performance: Faster integer conversion.
Because we know that the numbers we convert are valid integers and dont contain anything except the numericals and a minus sign, we can convert the byte array faster than int.Parse().

Saves some ~10% for both integer and string deserialization, depending on the exact inputs tested. Arrays and objects also benefit because of the length segment they have.